requirements.txt
fastapi
uvicorn
pillow
requests
numpy
safetensors
transformers
diffusers
torch
